+++ src/Components/Digital Card Website/Digital Card User Interface/Pages/DCLogin.js
@@ -5,6 +5,7 @@ import { useLocation, useNavigate } from 'react-router-dom';
 import Swal from 'sweetalert2';
 import { postData } from '../../../Services/NodeServices';
 import logo from "../../Digital Card Assets/IndiaBuzz.png";
+import logo1 from '../../Digital Card Assets/dchlogo.png';
 import bg from "../../Digital Card Assets/login_img.png";
 import Navbar from '../Components/Navbar';
 import OtpGenerator from '../ReviewTag/OtpGenerator';
@@ -33,38 +34,55 @@ export default function DCLogin() {
     // const handleClickShowPassword = () => setShowPassword(!showPassword);
     // const handleMouseDownPassword = () => setShowPassword(!showPassword);
     const handleSubmit = async () => {
-        var formData = new FormData()
+        if (phoneNo !== '') {
+            var formData = new FormData()
+            formData.append('phone', phoneNo)
+            var result = await postData('customerLogin/chkLogin', formData, true)
+            if (result.status) {
 
-        formData.append('phone', phoneNo)
-        // formData.append('password', password)
+                window.localStorage.setItem("userId", result.data._id)
+                window.localStorage.setItem("UserNumber", result?.data?.phone)
+                window.localStorage.setItem("UserEmail", result?.data?.email)
 
 
-        var result = await postData('customerLogin/chkLogin', formData, true)
-        
+                Swal.fire({
+                    title: 'Successfully Logged In!',
+                    imageUrl: logo1,
+                    imageWidth: 200,
+                    imageHeight: 200,
+                    imageAlt: 'Custom image',
+                    background: '#001e3c',
+                    timer: 1500,
+                    width: 500,
+                    padding: 15,
+                    color: '#fff',
+                    showConfirmButton: false,
 
-        if (result.status) {
+                })
+                navigate('/userdashboard')
+                window.localStorage.setItem("User", true)
+                window.localStorage.removeItem('data')
+                window.localStorage.setItem("data", JSON.stringify(result.data))
 
-            window.localStorage.setItem("userId", result.data._id)
-            window.localStorage.setItem("UserNumber", result?.data?.phone)
-            window.localStorage.setItem("UserEmail",  result?.data?.email)
-       
+            }
+            else {
+                Swal.fire({
+                    position: 'center',
+                    icon: 'error',
+                    title: 'Fail to Login',
+                    showConfirmButton: false,
+                    timer: 1500
+                })
 
-            enqueueSnackbar('Successfully Logged In!');
-            navigate('/userdashboard')
-            window.localStorage.setItem("User", true)
-            window.localStorage.removeItem('data')
-            window.localStorage.setItem("data", JSON.stringify(result.data))
-
-        }
-        else {
+            }
+        } else {
             Swal.fire({
                 position: 'center',
                 icon: 'error',
-                title: 'Fail to Login',
+                title: 'Enter Number First',
                 showConfirmButton: false,
                 timer: 1500
             })
-
         }
 
     }
@@ -185,17 +203,14 @@ export default function DCLogin() {
 
             setOtp(otpval)
 
-            const apiUrl = `https://soft7.in/api/send?number=91${phoneNo}&type=text&message=Your Otp For Digital Card Hub - ${otpval}&instance_id=65B92B5C6DD7D&access_token=65b928bbcea41`;
+            const apiUrl = `https://cloud.bulkpromo.in/api/send?number=91${phoneNo}&type=text&message=Your Otp For Digital Card Hub - ${otpval}&instance_id=6676AB42323B3&access_token=666ff52aa9a38`;
             const response = await postData('otp/api', { url: apiUrl })
-            // https://soft7.in/api/send?number=917225051627&type=text&message=test+message&instance_id=65B92B5C6DD7D&access_token=65b928bbcea41
         } else {
             Swal.fire({
                 text: "Enter the Number First",
                 timer: 1000
             })
         }
-
-        
     }
 
     // const dialogComponent = () => {
@@ -309,6 +324,7 @@ export default function DCLogin() {
                                 <Grid item xs={12}>
                                     {verified == true ? "Verified" : verified == false ? "Not Verified" : ""}
                                 </Grid>
+
                                 {/* <Grid item xs={12}>
                                     <TextField label="Password" type={showPassword ? "text" : "password"} // <-- This is where the magic happens
                                         InputProps={{ // <-- This is where the toggle button is added.
